requirements.txt
pre-commit

beautifulsoup4
selenium
pymongo
structlog
pydantic
pydantic-settings
aws-lambda-powertools

pika

qdrant-client
bytewax
unstructured
numpy
langchain
sentence-transformers
InstructorEmbedding

opik
langchain-openai
comet-ml
accelerate
bitsandbytes

sagemaker
